Move `Lockfile` into its own file in a new crate
So we can share it between snapshots and the CLI
