+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "crates/core",
   "crates/data-structures",
   "crates/durability",
+  "crates/fs-utils",
   "crates/lib",
   "crates/metrics",
   "crates/primitives",
@@ -98,6 +99,7 @@ spacetimedb-sats = { path = "crates/sats", version = "0.9.0" }
 spacetimedb-standalone = { path = "crates/standalone", version = "0.9.0" }
 spacetimedb-table = { path = "crates/table", version = "0.9.0" }
 spacetimedb-vm = { path = "crates/vm", version = "0.9.0" }
+spacetimedb-fs-utils = { path = "crates/fs-utils", version = "0.9.0" }
 
 ahash = "0.8"
 anyhow = "1.0.68"
