+++ crates/cli/Cargo.toml
@@ -24,6 +24,7 @@ spacetimedb-data-structures.workspace = true
 spacetimedb-lib.workspace = true
 spacetimedb-standalone = { workspace = true, optional = true }
 spacetimedb-client-api-messages.workspace = true
+spacetimedb-fs-utils.workspace = true
 
 anyhow.workspace = true
 base64.workspace = true
