+++ crates/cli/src/config.rs
@@ -3,6 +3,7 @@ use anyhow::Context;
 use jsonwebtoken::DecodingKey;
 use serde::{Deserialize, Serialize};
 use spacetimedb::auth::identity::decode_token;
+use spacetimedb_fs_utils::{create_parent_dir, lockfile::Lockfile};
 use spacetimedb_lib::Identity;
 use std::{
     fs,
@@ -103,65 +104,6 @@ pub struct RawConfig {
     identity_configs: Vec<IdentityConfig>,
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     server_configs: Vec<ServerConfig>,
-}
-
-fn create_parent_dir(file: &Path) -> anyhow::Result<()> {
-    let parent = file
-        .parent()
-        .with_context(|| format!("Cannot find the parent directory of path {file:?}"))?;
-
-    // If the `file` path is a relative path with no directory component,
-    // `parent` will be the empty path.
-    // In this case, do not attempt to create a directory.
-    if parent != Path::new("") {
-        // If the `file` path has a directory component,
-        // do `create_dir_all` to ensure it exists.
-        // If `parent` already exists as a directory, this is a no-op.
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("Failed to create directory structure {parent:?} to contain {file:?}"))?;
-    }
-    Ok(())
-}
-
-#[derive(Debug)]
-/// A file used as an exclusive lock on access to another file.
-///
-/// Constructing a `Lockfile` creates the `path` with [`std::fs::File::create_new`],
-/// a.k.a. `O_EXCL`, erroring if the file already exists.
-///
-/// Dropping a `Lockfile` deletes the `path`, releasing the lock.
-///
-/// Used to guarantee exclusive access to the system config file,
-/// in order to prevent racy concurrent modifications.
-struct Lockfile {
-    path: PathBuf,
-}
-
-impl Lockfile {
-    /// Acquire an exclusive lock on the configuration file `config_path`.
-    ///
-    /// `config_path` should be the full name of the SpacetimeDB configuration file.
-    fn for_config(config_path: &Path) -> anyhow::Result<Self> {
-        // Ensure the directory exists before attempting to create the lockfile.
-        create_parent_dir(config_path)?;
-
-        let mut path = config_path.to_path_buf();
-        path.set_extension("lock");
-        // Open with `create_new`, which fails if the file already exists.
-        std::fs::File::create_new(&path).with_context(|| {
-            "Unable to acquire lock on config file {config_path:?}: failed to create lockfile {path:?}"
-        })?;
-
-        Ok(Lockfile { path })
-    }
-}
-
-impl Drop for Lockfile {
-    fn drop(&mut self) {
-        std::fs::remove_file(&self.path)
-            .with_context(|| format!("Unable to remove lockfile {:?}", self.path))
-            .unwrap();
-    }
 }
 
 pub struct Config {
@@ -908,7 +850,7 @@ impl Config {
 
     pub fn load() -> Self {
         let home_path = Self::system_config_path();
-        let home_lock = Lockfile::for_config(&home_path).unwrap();
+        let home_lock = Lockfile::for_file(&home_path).unwrap();
         let home = if home_path.exists() {
             Self::load_from_file(&home_path)
                 .inspect_err(|e| eprintln!("config file {home_path:?} is invalid: {e:#?}"))
